+++ api/src/java/com/epam/esm/controller/TagController.java
@@ -1,10 +1,17 @@
 package com.epam.esm.controller;
 
 import com.epam.esm.dto.TagDTO;
+import com.epam.esm.exception.ControllerException;
 import com.epam.esm.exception.ServiceException;
 import com.epam.esm.service.TagService;
+import com.epam.esm.service.impl.GiftCertificateServiceImpl;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+
+import javax.validation.Valid;
 
 /**
  * Tag RestAPI.
@@ -13,11 +20,13 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/tag")
 public class TagController {
 
-    private TagService tagService;
+    private final TagService tagService;
 
     private TagController(TagService tagService) {
         this.tagService = tagService;
     }
+
+    private final static Logger logger = LogManager.getLogger(GiftCertificateServiceImpl.class);
 
     /**
      * Find tag by id.
@@ -25,11 +34,17 @@ public class TagController {
      * @param id the id
      * @return the response entity
      * @throws ServiceException the service exception
+     * @throws ControllerException if id is incorrect
      */
     @GetMapping("/{id}")
-    public ResponseEntity<TagDTO> find(@PathVariable(name = "id") Long id) throws ServiceException {
-        TagDTO tag = tagService.find(id);
-        return ResponseEntity.ok(tag);
+    public ResponseEntity<TagDTO> find(@PathVariable(name = "id") Long id) throws ServiceException, ControllerException {
+        if(id > 0 ){
+            TagDTO tag = tagService.find(id);
+            return ResponseEntity.ok(tag);
+        } else {
+            logger.error("Negative id exception");
+            throw  new ControllerException("Negative id exception");
+        }
     }
 
     /**
@@ -38,12 +53,15 @@ public class TagController {
      * @param tagName the tag name
      * @return the response entity
      * @throws ServiceException the service exception
+     * @throws ControllerException if tag name incorrect
      */
     @PostMapping(value = "/", consumes = "application/json")
-    public ResponseEntity<TagDTO> add(@RequestBody String tagName) throws ServiceException {
-        TagDTO tag = new TagDTO();
-        tag.setName(tagName);
-        TagDTO result = tagService.add(tag);
+    public ResponseEntity<TagDTO> add(@RequestBody @Valid TagDTO tagName, BindingResult bindingResult) throws ServiceException, ControllerException {
+        if (bindingResult.hasErrors()){
+            logger.error(bindingResultHandler(bindingResult));
+            throw new ControllerException(bindingResultHandler(bindingResult));
+        }
+        TagDTO result = tagService.add(tagName);
         return ResponseEntity.ok(result);
     }
 
@@ -53,10 +71,26 @@ public class TagController {
      * @param id the id
      * @return the response entity
      * @throws ServiceException the service exception
+     * @throws ControllerException if id is incorrect
      */
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> delete(@PathVariable(name = "id") Long id) throws ServiceException {
-        long result = tagService.delete(id);
-        return result != -1L ? ResponseEntity.ok("Delete successful!") : ResponseEntity.ok("Delete unsuccessful!");
+    public ResponseEntity<String> delete(@PathVariable(name = "id") Long id) throws ServiceException, ControllerException {
+        if (id > 0) {
+            long result = tagService.delete(id);
+            return result != -1L ? ResponseEntity.ok("Delete successful!") : ResponseEntity.ok("Delete unsuccessful!");
+
+        } else {
+            logger.error("Negative id exception");
+            throw  new ControllerException("Negative id exception");
+        }
+    }
+
+    /**
+     * Get default message by validate exception
+     * @param bindingResult exceptions of validate
+     * @return string default message of exception
+     */
+    private String bindingResultHandler(BindingResult bindingResult){
+       return bindingResult.getAllErrors().get(0).getDefaultMessage();
     }
 }
